backend/pyproject.toml
[project]
name = "stream-prompts-api"
version = "0.1.0"
description = "FastAPI backend for Stream Prompts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
]

[project.scripts]
app = "app.main:app"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.vercel.scripts]
build = ""

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
